fix #3588: helper: consider NIC count for config-specific timeout
There have been some reports about `qm start` timeouts on VMs that have a
lot of NICs assigned.
This patch considers the number of NICs when calculating the config-specific
timeout. Since the increase in start time is linearly related to the number
of NICs, a constant timeout increment per NIC was chosen.

Signed-off-by: Hannes Laimer <h.laimer@proxmox.com>