drive: add helpers for additional checks for storage layer info about volume
First step towards using the storage layer format instead of the
extension based format from qemu_img_format() as a source of truth
everywhere. Currently, some callers use qemu_img_format() and some
use parse_volname().

For import, special handling is needed, because the format can be a
combined ova+$extracted_format.

There is a fallback for 'raw' format when no format is returned by the
storage layer for backwards compatibility, e.g. ISOs. Formats that are
not part of the $QEMU_FORMAT_RE are not allowed.

Signed-off-by: Fiona Ebner <f.ebner@proxmox.com>