vm start: log task warning if both SEV-SNP and efidisk are configured
Make it more noticeable if a efidisk will be ignored, as this means
one changed a existing VM that already had a efidisk entry to SEV-SNP,
as adding an efidisk for such VMs with SEV-SNP enabled is blocked.

While at it fix indentation and drop the duplicated $arch variable,
that's already declared as required here at the top of the method.

Signed-off-by: Thomas Lamprecht <t.lamprecht@proxmox.com>