tests: cfg2cmd: mock get_hw_capabilities to allow testing SEV
Allow tests to declare either a raw HW capabilities string as our
minitool produces or refer to a predefined CPU series, for now that's
only 'amd-turin-9005' which I pulled of a modern AMD EPYC 9475F based
host.

If nothing is set then default to a very barebones HW capability set.

Signed-off-by: Thomas Lamprecht <t.lamprecht@proxmox.com>